+++ create-vite-app/template-react/vite.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  alias: {
-    react: '@pika/react',
-    'react-dom': '@pika/react-dom'
-  },
-  jsx: 'react'
-}