+++ src/node/config.ts
@@ -8,9 +8,12 @@ import { Options as RollupPluginVueOptions } from 'rollup-plugin-vue'
 import { CompilerOptions } from '@vue/compiler-sfc'
 import Rollup, {
   InputOptions as RollupInputOptions,
-  OutputOptions as RollupOutputOptions
+  OutputOptions as RollupOutputOptions,
+  OutputChunk
 } from 'rollup'
 import { Transform } from './transform'
+import { DepOptimizationOptions } from './depOptimizer'
+import { IKoaProxiesOptions } from 'koa-proxies'
 
 export { Resolver, Transform }
 
@@ -26,6 +29,16 @@ export interface SharedConfig {
   root?: string
   /**
    * Import alias. Can only be exact mapping, does not support wildcard syntax.
+   *
+   * Example `vite.config.js`:
+   * ``` js
+   * module.exports = {
+   *   alias: {
+   *     'react': '@pika/react',
+   *     'react-dom': '@pika/react-dom'
+   *   }
+   * }
+   * ```
    */
   alias?: Record<string, string>
   /**
@@ -38,16 +51,27 @@ export interface SharedConfig {
    */
   resolvers?: Resolver[]
   /**
-   * Options to pass to @vue/compiler-dom
+   * Configure dep optimization behavior.
+   *
+   * Example `vite.config.js`:
+   * ``` js
+   * module.exports = {
+   *   optimizeDeps: {
+   *     exclude: ['dep-a', 'dep-b']
+   *   }
+   * }
+   * ```
+   */
+  optimizeDeps?: DepOptimizationOptions
+  /**
+   * Options to pass to `@vue/compiler-dom`
+   *
+   * https://github.com/vuejs/vue-next/blob/master/packages/compiler-core/src/options.ts
    */
   vueCompilerOptions?: CompilerOptions
   /**
    * Configure what to use for jsx factory and fragment.
-   * @default
-   * {
-   *   factory: 'React.createElement',
-   *   fragment: 'React.Fragment'
-   * }
+   * @default 'vue'
    */
   jsx?:
     | 'vue'
@@ -61,6 +85,32 @@ export interface SharedConfig {
 
 export interface ServerConfig extends SharedConfig {
   /**
+   * Configure custom proxy rules for the dev server. Uses
+   * [`koa-proxies`](https://github.com/vagusX/koa-proxies) which in turn uses
+   * [`http-proxy`](https://github.com/http-party/node-http-proxy). Each key can
+   * be a path Full options
+   * [here](https://github.com/http-party/node-http-proxy#options).
+   *
+   * Example `vite.config.js`:
+   * ``` js
+   * module.exports = {
+   *   proxy: {
+   *     proxy: {
+   *       // string shorthand
+   *       '/foo': 'http://localhost:4567/foo',
+   *       // with options
+   *       '/api': {
+   *         target: 'http://jsonplaceholder.typicode.com',
+   *         changeOrigin: true,
+   *         rewrite: path => path.replace(/^\/api/, '')
+   *       }
+   *     }
+   *   }
+   * }
+   * ```
+   */
+  proxy?: Record<string, string | IKoaProxiesOptions>
+  /**
    * Whether to use a Service Worker to cache served code. This can greatly
    * improve full page reload performance, but requires a Service Worker
    * update + reload on each server restart.
@@ -68,7 +118,7 @@ export interface ServerConfig extends SharedConfig {
    * @default false
    */
   serviceWorker?: boolean
-  plugins?: ServerPlugin[]
+  configureServer?: ServerPlugin | ServerPlugin[]
 }
 
 export interface BuildConfig extends SharedConfig {
@@ -86,7 +136,7 @@ export interface BuildConfig extends SharedConfig {
   /**
    * Directory relative from `outDir` where the built js/css/image assets will
    * be placed.
-   * @default 'assets'
+   * @default '_assets'
    */
   assetsDir?: string
   /**
@@ -115,16 +165,19 @@ export interface BuildConfig extends SharedConfig {
   // The following are API only and not documented in the CLI. -----------------
   /**
    * Will be passed to rollup.rollup()
+   *
    * https://rollupjs.org/guide/en/#big-list-of-options
    */
   rollupInputOptions?: RollupInputOptions
   /**
    * Will be passed to bundle.generate()
+   *
    * https://rollupjs.org/guide/en/#big-list-of-options
    */
   rollupOutputOptions?: RollupOutputOptions
   /**
    * Will be passed to rollup-plugin-vue
+   *
    * https://github.com/vuejs/rollup-plugin-vue/blob/next/src/index.ts
    */
   rollupPluginVueOptions?: Partial<RollupPluginVueOptions>
@@ -148,13 +201,15 @@ export interface BuildConfig extends SharedConfig {
    * @default true
    */
   emitAssets?: boolean
+  /**
+   * Predicate function that determines whether a link rel=modulepreload shall be
+   * added to the index.html for the chunk passed in
+   */
+  shouldPreload?: (chunk: OutputChunk) => boolean
 }
 
-export interface UserConfig
-  extends BuildConfig,
-    Pick<ServerConfig, 'serviceWorker'> {
+export interface UserConfig extends BuildConfig, ServerConfig {
   plugins?: Plugin[]
-  configureServer?: ServerPlugin
 }
 
 export interface Plugin
@@ -246,9 +301,6 @@ export async function resolveConfig(
       for (const plugin of config.plugins) {
         config = resolvePlugin(config, plugin)
       }
-      // delete plugins so it doesn't get passed to `createServer` as server
-      // plugins.
-      delete config.plugins
     }
 
     require('debug')('vite:config')(
@@ -299,14 +351,10 @@ function resolvePlugin(config: UserConfig, plugin: Plugin): UserConfig {
     },
     transforms: [...(config.transforms || []), ...(plugin.transforms || [])],
     resolvers: [...(config.resolvers || []), ...(plugin.resolvers || [])],
-    configureServer: (ctx) => {
-      if (config.configureServer) {
-        config.configureServer(ctx)
-      }
-      if (plugin.configureServer) {
-        plugin.configureServer(ctx)
-      }
-    },
+    configureServer: ([] as any[]).concat(
+      config.configureServer || [],
+      plugin.configureServer || []
+    ),
     vueCompilerOptions: {
       ...config.vueCompilerOptions,
       ...plugin.vueCompilerOptions