+++ src/node/depOptimizer.ts
@@ -3,20 +3,53 @@ import path from 'path'
 import { createHash } from 'crypto'
 import { ResolvedConfig } from './config'
 import type Rollup from 'rollup'
-import { createResolver, resolveNodeModule } from './resolver'
+import {
+  createResolver,
+  supportedExts,
+  resolveNodeModuleEntry
+} from './resolver'
 import { createBaseRollupPlugins } from './build'
-import { resolveFrom } from './utils'
+import { resolveFrom, lookupFile } from './utils'
 import { init, parse } from 'es-module-lexer'
 import chalk from 'chalk'
 import { Ora } from 'ora'
+import { createBuildCssPlugin } from './build/buildPluginCss'
+
+const KNOWN_IGNORE_LIST = new Set([
+  'tailwindcss',
+  '@tailwindcss/ui',
+  '@pika/react',
+  '@pika/react-dom'
+])
+
+export interface DepOptimizationOptions {
+  /**
+   * Only optimize explicitly listed dependencies.
+   */
+  include?: string[]
+  /**
+   * Do not optimize these dependencies.
+   */
+  exclude?: string[]
+  /**
+   * Explicitly allow these CommonJS deps to be bundled.
+   */
+  commonJSWhitelist?: string[]
+  /**
+   * Automatically run `vite optimize` on server start?
+   * @default true
+   */
+  auto?: boolean
+}
 
 export const OPTIMIZE_CACHE_DIR = `node_modules/.vite_opt_cache`
 
-export interface OptimizeOptions extends ResolvedConfig {
-  force?: boolean
-}
-
-export async function optimizeDeps(config: OptimizeOptions, asCommand = false) {
+export async function optimizeDeps(
+  config: ResolvedConfig & { force?: boolean },
+  asCommand = false
+) {
+  const debug = require('debug')('vite:optimize')
+  const log = asCommand ? console.log : debug
   const root = config.root || process.cwd()
   // warn presence of web_modules
   if (fs.existsSync(path.join(root, 'web_modules'))) {
@@ -28,7 +61,13 @@ export async function optimizeDeps(config: OptimizeOptions, asCommand = false) {
     )
   }
 
-  const cacheDir = path.join(root, OPTIMIZE_CACHE_DIR)
+  const pkgPath = lookupFile(root, [`package.json`], true /* pathOnly */)
+  if (!pkgPath) {
+    log(`package.json not found. Skipping.`)
+    return
+  }
+
+  const cacheDir = path.join(path.dirname(pkgPath), OPTIMIZE_CACHE_DIR)
   const hashPath = path.join(cacheDir, 'hash')
   const depHash = getDepHash(root, config.__path)
 
@@ -39,8 +78,7 @@ export async function optimizeDeps(config: OptimizeOptions, asCommand = false) {
     } catch (e) {}
     // hash is consistent, no need to re-bundle
     if (prevhash === depHash) {
-      asCommand &&
-        console.log('Hash is consistent. Skipping. Use --force to override.')
+      log('Hash is consistent. Skipping. Use --force to override.')
       return
     }
   }
@@ -48,51 +86,99 @@ export async function optimizeDeps(config: OptimizeOptions, asCommand = false) {
   await fs.remove(cacheDir)
   await fs.ensureDir(cacheDir)
 
-  const pkg = lookupFile(root, [`package.json`])
-  if (!pkg) {
-    asCommand && console.log(`package.json not found. Skipping.`)
-    return
-  }
-
-  const deps = Object.keys(JSON.parse(pkg).dependencies || {})
+  const deps = Object.keys(require(pkgPath).dependencies || {})
   if (!deps.length) {
-    asCommand &&
-      console.log(`No dependencies listed in package.json. Skipping.`)
+    await fs.writeFile(hashPath, depHash)
+    log(`No dependencies listed in package.json. Skipping.`)
     return
   }
 
   const resolver = createResolver(root, config.resolvers, config.alias)
+  const { include, exclude, commonJSWhitelist } = config.optimizeDeps || {}
 
   // Determine deps to optimize. The goal is to only pre-bundle deps that falls
   // under one of the following categories:
-  // 1. Is CommonJS module
-  // 2. Has imports to relative files (e.g. lodash-es, lit-html)
-  // 3. Has imports to bare modules that are not in the project's own deps
+  // 1. Has imports to relative files (e.g. lodash-es, lit-html)
+  // 2. Has imports to bare modules that are not in the project's own deps
   //    (i.e. esm that imports its own dependencies, e.g. styled-components)
   await init
+  const cjsDeps: string[] = []
   const qualifiedDeps = deps.filter((id) => {
-    const entry = resolveNodeModule(root, id)
-    if (!entry) {
+    if (include && !include.includes(id)) {
+      debug(`skipping ${id} (not included)`)
+      return false
+    }
+    if (exclude && exclude.includes(id)) {
+      debug(`skipping ${id} (excluded)`)
+      return false
+    }
+    if (commonJSWhitelist && commonJSWhitelist.includes(id)) {
+      debug(`optimizing ${id} (commonJSWhitelist)`)
+      return true
+    }
+    if (KNOWN_IGNORE_LIST.has(id)) {
+      debug(`skipping ${id} (internal excluded)`)
+      return false
+    }
+    const pkgInfo = resolveNodeModuleEntry(root, id)
+    if (!pkgInfo) {
+      debug(`skipping ${id} (cannot resolve entry)`)
+      return false
+    }
+    const [entry, pkg] = pkgInfo
+    if (!supportedExts.includes(path.extname(entry))) {
+      debug(`skipping ${id} (entry is not js)`)
       return false
     }
     const content = fs.readFileSync(resolveFrom(root, entry), 'utf-8')
     const [imports, exports] = parse(content)
-    if (!exports.length) {
-      // no exports, likely a commonjs module
-      return true
+    if (!exports.length && !/export\s+\*\s+from/.test(content)) {
+      if (!pkg.module) {
+        cjsDeps.push(id)
+      }
+      debug(`skipping ${id} (no exports, likely commonjs)`)
+      return false
     }
-    //
     for (const { s, e } of imports) {
       let i = content.slice(s, e).trim()
       i = resolver.alias(i) || i
-      if (i.startsWith('.') || !deps.includes(i)) {
+      if (i.startsWith('.')) {
+        debug(`optimizing ${id} (contains relative imports)`)
+        return true
+      }
+      if (!deps.includes(i)) {
+        debug(`optimizing ${id} (imports sub dependencies)`)
         return true
       }
     }
+    debug(`skipping ${id} (single esm file, doesn't need optimization)`)
   })
 
   if (!qualifiedDeps.length) {
-    asCommand && console.log(`No listed dependency requires optimization.`)
+    if (!cjsDeps.length) {
+      await fs.writeFile(hashPath, depHash)
+      log(`No listed dependency requires optimization. Skipping.`)
+    } else {
+      console.error(
+        chalk.yellow(
+          `[vite] The following dependencies seem to be CommonJS modules that\n` +
+            `do not provide ESM-friendly file formats:\n\n  ` +
+            cjsDeps.map((dep) => chalk.magenta(dep)).join(`\n  `) +
+            `\n` +
+            `\n- If you are not using them in browser code, you can move them\n` +
+            `to devDependencies or exclude them from this check by adding\n` +
+            `them to ${chalk.cyan(
+              `optimizeDeps.exclude`
+            )} in vue.config.js.\n` +
+            `\n- If you do intend to use them in the browser, you can try adding\n` +
+            `them to ${chalk.cyan(
+              `optimizeDeps.commonJSWhitelist`
+            )} in vue.config.js but they\n` +
+            `may fail to bundle or work properly. Consider choosing more modern\n` +
+            `alternatives that provide ES module build formts.`
+        )
+      )
+    }
     return
   }
 
@@ -106,7 +192,7 @@ export async function optimizeDeps(config: OptimizeOptions, asCommand = false) {
   const msg = asCommand
     ? `Pre-bundling dependencies to speed up dev server page load...`
     : `Pre-bundling them to speed up dev server page load...\n` +
-      `  (this will be run only when your dependencies have changed)`
+      `(this will be run only when your dependencies have changed)`
   if (process.env.DEBUG || process.env.NODE_ENV === 'test') {
     console.log(msg)
   } else {
@@ -124,17 +210,21 @@ export async function optimizeDeps(config: OptimizeOptions, asCommand = false) {
     }, {} as Record<string, string>)
 
     const rollup = require('rollup') as typeof Rollup
+    const warningIgnoreList = [`CIRCULAR_DEPENDENCY`, `THIS_IS_UNDEFINED`]
     const bundle = await rollup.rollup({
       input,
       external: preservedDeps,
       treeshake: { moduleSideEffects: 'no-external' },
       onwarn(warning, warn) {
-        if (warning.code !== 'CIRCULAR_DEPENDENCY') {
+        if (!warningIgnoreList.includes(warning.code!)) {
           warn(warning)
         }
       },
       ...config.rollupInputOptions,
-      plugins: await createBaseRollupPlugins(root, resolver, config)
+      plugins: [
+        ...(await createBaseRollupPlugins(root, resolver, config)),
+        createBuildCssPlugin(root, '/', 'assets')
+      ]
     })
 
     const { output } = await bundle.generate({
@@ -146,19 +236,24 @@ export async function optimizeDeps(config: OptimizeOptions, asCommand = false) {
 
     spinner && spinner.stop()
 
+    const optimized = []
     for (const chunk of output) {
       if (chunk.type === 'chunk') {
         const fileName = chunk.fileName
         const filePath = path.join(cacheDir, fileName)
         await fs.ensureDir(path.dirname(filePath))
         await fs.writeFile(filePath, chunk.code)
-        console.log(
-          `${chalk.yellow(fileName.replace(/\.js$/, ''))} -> ${chalk.dim(
-            path.relative(root, filePath)
-          )}`
-        )
+        if (!fileName.startsWith('common/')) {
+          optimized.push(fileName.replace(/\.js$/, ''))
+        }
       }
     }
+
+    console.log(
+      `Optimized modules:\n${optimized
+        .map((id) => chalk.yellowBright(id))
+        .join(`, `)}`
+    )
 
     await fs.writeFile(hashPath, depHash)
   } catch (e) {
@@ -166,7 +261,15 @@ export async function optimizeDeps(config: OptimizeOptions, asCommand = false) {
     if (asCommand) {
       throw e
     } else {
-      console.error(`[vite] `)
+      console.error(chalk.red(`[vite] Dep optimization failed with error:`))
+      console.error(e)
+      console.log()
+      console.log(
+        chalk.yellow(
+          `Tip: You can configure what deps to include/exclude for optimization\n` +
+            `using the \`optimizeDeps\` option in the Vite config file.`
+        )
+      )
     }
   }
 }
@@ -183,23 +286,11 @@ export function getDepHash(
     return cachedHash
   }
   let content = lookupFile(root, lockfileFormats) || ''
-  content += lookupFile(root, [`package.json`]) || ''
+  const pkg = JSON.parse(lookupFile(root, [`package.json`]) || '{}')
+  content += JSON.stringify(pkg.dependencies)
   // also take config into account
   if (configPath) {
     content += fs.readFileSync(configPath, 'utf-8')
   }
   return createHash('sha1').update(content).digest('base64')
-}
-
-function lookupFile(dir: string, formats: string[]): string | undefined {
-  for (const format of formats) {
-    const fullPath = path.join(dir, format)
-    if (fs.existsSync(fullPath)) {
-      return fs.readFileSync(fullPath, 'utf-8')
-    }
-  }
-  const parentDir = path.dirname(dir)
-  if (parentDir !== dir) {
-    return lookupFile(parentDir, formats)
-  }
 }