+++ src/node/utils/fsUtils.ts
@@ -84,3 +84,20 @@ export async function readBody(
     return !stream || typeof stream === 'string' ? stream : stream.toString()
   }
 }
+
+export function lookupFile(
+  dir: string,
+  formats: string[],
+  pathOnly = false
+): string | undefined {
+  for (const format of formats) {
+    const fullPath = path.join(dir, format)
+    if (fs.existsSync(fullPath)) {
+      return pathOnly ? fullPath : fs.readFileSync(fullPath, 'utf-8')
+    }
+  }
+  const parentDir = path.dirname(dir)
+  if (parentDir !== dir) {
+    return lookupFile(parentDir, formats, pathOnly)
+  }
+}